Change StringExtension to SpotifyClientExtension As mentioned in the pull request, changing the string extension to a Spotify Client extensions feels better, at least imo. So I've done that.
